Refactored Dataset class
